Added ArrayFunctionHolder class & StringFunctionHolder class to organize algorithms
