--- DataStructuresAndAlgorithimsV2/Tweet.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DataStructuresAndAlgorithimsV2
-{
-    class Tweet
-    {
-        public string Message { get; set; }
-        public int Date { get; set; }
-    }
-}
